requirements.txt
asgiref==3.4.1
design-django-theme==1.4.6
Django==3.2.8
django-ajax-datatable==4.3.1
django-appconf==1.0.5
django-compressor==2.4.1
django-sass-processor==1.1
django-tables2==2.4.1
djangorestframework==3.12.4
libsass==0.21.0
mysqlclient==2.0.3
protobuf==3.18.1
pytz==2021.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
